Fixes bug in cpu_relax_vc()
It would never ensure the other vcores run, since spun would be reset on
every invocation.

Note that since it is a TLS variable, it gets reused each time and never
get reset.  So if a spinner calls cpu_relax_vc() 100 times before
grabbing its lock (or whatever), then the next time that vcore relaxes,
it'll start at 100.  Only 900 to go!

I'm okay with this - it's a little easier to deal with than having the
caller pass in the spin variable.