Fixes handle_indirs issue
handle_indirs() can only proceed if the target vcore is guaranteed to
not be online.  VC_PREEMPTED is the commonly used flag for that, with
the added benefit of that flag turning off the first time a VC comes
back online - meaning we can skip handling their messages if its a moot
point.

This change sets PREEMPTED for all cases of change_to_vcore, since the
"I don't plan to return" case needs to have its messages checked
remotely.