Fixes bug with EVENT_VCORE_MUST_RUN
Spamming list members would succeed for offline vcores, which breaks if
MUST_RUN was set.  The race was that a VC could have been on the online
list, but by the time we messaged it, it was preempted (and it still
can_rcv).  Likewise, checking the bulk-preempt and inactive lists could
be racy as well.

Four bugs found while writing a dissertation.  Not too bad.  =)