Increases the op backtrace buffer
Our callgraphs are often deeper than 8, which leads to not properly attributing
deep functions to their older ancestors.  The downside is the buffers (on the
stack) are bigger, increasing our chances of running off the stack, and
increasing the overhead of the backtrace.  If the stack is a problem, we can
malloc the buffer or something.