Add a regression device.
The regression device is used for regression tests. It is '#Z'
and provides mondata and monctl.

monctl is currently not useful.

mondata on the write side takes monitor commands.

Currently on the read side it returns nothing, but soon
it will return the output of the monitor for that command.

Sample usage:
/ $ echo ps > '#Z/mondata'
     PID Name                 State      Parent
-------------------------------------------------
       1 busybox              WAITING         0
       2 echo                 RUNNING_S       1

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>