Add define for _GLIBCXX_NO_IOCTL
The implementation of std::streambuf::showmanyc in
libstdc++-v3/src/c++98/basic_file.cc, uses various methods to estimate
on the number of characters available in the associated input sequence.
Using ioctls is one of them.  We don't support ioctls on Akaros, though,
so we resort to one of the other methods (of which reutning 0 is OK).