__ctype_init() every new TLS (XCC)
Functions like isspace() look in TLS for a pointer to an array used in the
ctype ops.  If you PF near there, then that pointer might not be set.
__ctype_init() initializes that pointer.

Every TLS needs this done once.  We do it a little more often than necessary
(right before vcore_entry, for example).  Thread0 also had a few issues with
this.  The new init-array business in glibc doesn't call _init (hence the
override of elf-init.c), and our custom init method didn't catch the update to
call __ctype_init().