Adds openat() stubs (XCC)
Glibc now implements the POSIX/UNIX opendir() with openat().  The kernel
doesn't support openat() yet, but we can still use the interface for absolute
or AT_FDCWD as a wrapper around open().  When we do support it, the utest ought
to pass.

This fixes things like 'ls' and tab completion in busybox (which needs to do an
ls).