Adds __ALIGN_MASK (XCC)
Helper used in bitmap code.  We could use it in kmalloc_align if we want.  To
keep that option open, I've asserted kmalloc aligns are a power of 2.  The
current code (ROUNDUP) can handle non-powers of 2.

Technically, ros/common.h is a kernel header, but don't worry about it in this
case.