Implements werrstr in glibc (XCC)
Userspace needs a helper to write errstr.  The kernel's version is
set_errstr().  I left the user's as werrstr, which is the Plan 9 name for the
function.

I put it in glibc, since that manages the errstr code and also uses it (e.g.
openat()).  Finding a place for it was a pain.  Don't try to #include stdio.h
from within errno.c.

Incidentally, the #warning to do this was around a while, but it was a bit of
motivation to get it done (esp as part of the new toolchain).