Adds stub for gethostname (XCC)
One of the changes in glibc from 2.14 to 2.19 was the inclusion of
sysdeps/posix/gethostname.c, which overrides the old misc/gethostname.c.  The
new one calls uname() (from posix/uname.c), which calls gethostname.c.  That
infinite loop eventually runs out off the end of the stack, causing a page
fault in netperf.

In lieu of building in gethostname (which I imagine could be a file in devmisc
or something), we just error out like we did in glibc 2.14.

Copy the new file into glibc-2.19 and rebuild, or just make clean.