Adds interfaces to get/put CG cores for the kernel
For things like the ARSCs, NIX mode, or whatever else you want.  This takes a
core out of the idle list, meaning MCPs will not receive them.

If a core was idle (unallocated) but provisioned, these interfaces will not
choose those cores.  It would mess up other parts of the ksched.