BSD taskqueues via KMSGs
BSD taskqueues are kernel threads that execute work asynchronously.  Check out
http://www.freebsd.org/cgi/man.cgi?query=taskqueue for more info.

We could implement them with ktasks that sleep on queues, and pump the tasks
into the queues.  But the tasks are basically slimmed down kernel messages, so
we can just send a message.

In the future, we may want better support or more taskqueue behaviors, both in
BSD code and more generally throughout Akaros.  We can always reimplement
taskqueues in a more appropriate manner at that time.