Bus space barriers (XCC)
On x86, a LOCK will flush the write buffers.  This is what BSD does under the
hood for their bus_space_barriers.  I don't know when the LOCKs should be used
compared to lfence/sfence/mfence.

Technically, these are kernel headers we're messing with (the membar), but
don't worry about it.