BXE: basic probe
Then it bails out.

One difference between Plan 9 and BSD is when the ctlr gets allocated.  Plan 9
does it in the PCI/PNP stage.  In BSD, the sizeof (or something) is declared
earlier and someone else does it.  I think.  I'm less a fan of the latter part;
maybe I don't understand it.  But the pcidev->some_void_star = sc might be
useful.