BXE: Manages if 0 a bit
I went through the original driver and made sure anything that we freshly
marked if 0 is marked like this:

 #if 0 // AKAKROS_PORT

Using "if 0" is useful at least for me, since my editor syntax-highlights
everything in an if 0 the color of a comment - meaning I don't look at it.  A
regular ifdef wouldn't do that.

I also uncommented a few more things that compile, getting us a little closer
to initializing the NIC.