BNX2X: IO funcs spatched
e.g. readl -> read32.

Yes, we could have kept the Linux io helpers the way Linux expects, but if
we're putting something useful in the main kernel, I'd like it to be the way we
want.

Though the spatch did require --disable-multi-pass for it to work on the macros
in bnx2x.h, and the writel macros needed to be manually joined onto one line.
I think those were broken by a previous spatch.