BNX2X: spatch net_device -> ether
The fields aren't right, but it's the closest thing.  And at least the name
field is right!  (Thanks to the anonymous netif in ether).

Incidentally, spatch couldn't handle two of the changes.  Again, one in bnx2x.h
and another in bnx2x_main.c.  In the latter, if you remove the "return -EAGAIN"
at 12208, it works...