BNX2X: spatch netif_addr_lock
It's not clear that the plan 9 qlock is the right thing to grab here.
It is the equivalent pattern, I think, but it's a qlock instead of a
spinlock.  There might be problems with grabbing a qlock in the wrong
context (IRQ, etc).

An alternative is to put a spinlock in plan 9 and use that.  Maybe they
don't need a qlock.