BNX2X: spatch netdevice feature conversion
Spatch missed kern/drivers/net/bnx2x/bnx2x_cmn.c.

Note that this removes the hw_enc and vlan features, and #define's the
hw_features to the normal features.

I left the Linux NETIF_F_ flags in place, and just #defined the ones we support
for now.  There might be issues with #defining them to 0, in case there's some
code that assumes a flag #define value is non-zero (consider the issues with
openmode and O_RDONLY being 0).