Adds ALIGN and PAGE_ALIGN (XCC)
Slightly modified versions of Linux's macros, to minimize pointer issues.

ALIGN can't be exposed to userspace.  There is another one of those floating
around.  The reason we had things in common.h was that we wanted to use them in
user/ without writing them down twice.  As we add items, we're more likely to
run into a conflict with glibc.  Maybe the whole thing needs reworked.

Technically, this changed a kernel header, though it should only affect the
kernel.