Scaffolded '#T' ttrace device.
This commit has all of the plan 9 bits to implement the #T device. It
does not contain any code that can take or report samples yet. But it
has all of the handling for writing to the cpu data files and the
9p layout of all of the files that will report data.

Has some scaffolding which #warns during compilation.

Started on some simple tests/ttrace.c which can read and decode the data
file such as it is currently.