Fix kprof attach leaks.
When you perform 'ls -lR \#K' on a system that hasn't bound the kprof
device it repeatedly attaches/closes the #K device. This leaks
cpu_buffers (not a big issue, alloc_cpu_buffers defends itself),
oprof_alarms and kprof.systrace queue.

This fix is not ideal, but it only allocates the resources once, I
didn't want to implement an atomic execute once when I don't understand
the code in detail.